@zephraph => [Redesign] Update pages for display in force
