+++ src/Styleguide/Components/Slider.tsx
@@ -99,6 +99,10 @@ const ImageContainer = styled.div.attrs<ImageContainerProps>({})`
   width: 100%;
   height: 100%;
 
+  img {
+    display: inline-block;
+  }
+
   ${width};
   ${height};
   ${border};
