+++ src/Styleguide/Components/Tabs.tsx
@@ -107,6 +107,7 @@ const TabContainer = styled.div`
   padding-bottom: 13px;
   margin-bottom: -1px;
   margin-right: 20px;
+  white-space: nowrap;
   ${borders};
 `
