+++ src/Styleguide/Pages/Artist/Overview.js
@@ -1,4 +1,4 @@
-import { Sans } from "@artsy/palette"
+import { Sans, Serif } from "@artsy/palette"
 import React from "react"
 import styled from "styled-components"
 import { space, width } from "styled-system"
@@ -66,10 +66,12 @@ export const Overview = () => {
               const geneDivider = index < list.length - 1 ? "," : ""
 
               return (
-                <GeneFamilyItem href="#" className="noUnderline" mr={0.5}>
-                  {gene}
-                  {geneDivider}
-                </GeneFamilyItem>
+                <Serif size="3t" display="inline-block" key={index}>
+                  <GeneLink href="#" className="noUnderline" mr={0.5}>
+                    {gene}
+                    {geneDivider}
+                  </GeneLink>
+                </Serif>
               )
             })}
           </GeneFamily>
@@ -147,7 +149,7 @@ export const Overview = () => {
 }
 
 const GeneFamily = styled.div``
-const GeneFamilyItem = styled.a`
+const GeneLink = styled.a`
   display: inline-block;
   ${space};
 `
