+++ src/Styleguide/Pages/Fixtures/Artwork/imageData.ts
@@ -1,0 +1,8 @@
+export function imageData(width: number, height: number, imageIndex?: number) {
+  return {
+    uri: `https://picsum.photos/${width}/${height}/?${
+      imageIndex === undefined ? "random" : `image=${imageIndex}`
+    }`,
+    aspectRatio: width / height,
+  }
+}
